[新增功能](master): 分页参数统一修改为long型.

+++ src/main/java/com/catface/common/model/OffsetPageModel.java
@@ -1,10 +1,10 @@
 package com.catface.common.model;
 
-import java.util.List;
-
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
+
+import java.util.List;
 
 /**
  * @author by catface
@@ -25,12 +25,12 @@ public class OffsetPageModel<O, T> {
     private O maxOffset;
 
     @ApiModelProperty(value = "分页大小")
-    private Integer size;
+    private long size;
 
     @ApiModelProperty(value = "是否有下一页,true:有;false:无")
     private Boolean hasNext;
 
-    public OffsetPageModel(List<T> records, O minOffset, O maxOffset, Integer size, boolean hasNext) {
+    public OffsetPageModel(List<T> records, O minOffset, O maxOffset, long size, boolean hasNext) {
         this.records = records;
         this.minOffset = minOffset;
         this.maxOffset = maxOffset;
