+++ src/main/java/com/catface/common/model/OffsetParam.java
@@ -1,11 +1,11 @@
 package com.catface.common.model;
 
-import javax.validation.constraints.Min;
-
 import com.catface.common.enums.common.QueryDirectionEnum;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
+
+import javax.validation.constraints.Min;
 
 /**
  * @author by catface
@@ -23,13 +23,13 @@ public class OffsetParam<O> {
 
     @ApiModelProperty(value = "分页大小")
     @Min(value = 1, message = "分页大小需大于0")
-    private Integer size = 10;
+    private long size = 10L;
 
     public OffsetParam() {
 
     }
 
-    public OffsetParam(O offset, Integer size, QueryDirectionEnum direction) {
+    public OffsetParam(O offset, Long size, QueryDirectionEnum direction) {
         this.offset = offset;
         this.size = size;
         this.direction = direction;
